[ADD] IpReleasedEventListener 에서 handling 로직 분리
