+++ src/main/kotlin/com/iplease/server/ip/manage/domain/assign/handler/IpAssignedEventHandler.kt
@@ -1,0 +1,7 @@
+package com.iplease.server.ip.manage.domain.assign.handler
+
+import com.iplease.server.ip.manage.global.common.data.dto.AssignedIpDto
+
+interface IpAssignedEventHandler {
+    fun handle(dto: AssignedIpDto)
+}
